refactor: extract UserDialogDelegate from EntityImp (786→400 lines)
Extract all user-prompt dialog UI code (JNumPad, NumberModel hierarchy,
and the four getXxxFromUser methods) into a package-private
UserDialogDelegate class. EntityImp retains thin delegation methods so
its public API is unchanged.

- EntityImp: 786 → 400 lines (target was <500)
- New file: UserDialogDelegate.java (package-private)
- New test: EntityImpStructureTest.java (API surface, visibility, signatures)
- All 337 Maven tests pass; all 755 Python contract tests pass

Closes #601
