+++ core/story-api/src/test/java/org/lgna/story/implementation/EntityImpStructureTest.java
@@ -1,0 +1,133 @@
+package org.lgna.story.implementation;
+
+import org.junit.Test;
+
+import java.lang.reflect.Method;
+import java.lang.reflect.Modifier;
+import java.util.Arrays;
+import java.util.Set;
+import java.util.TreeSet;
+import java.util.stream.Collectors;
+
+import static org.junit.Assert.*;
+
+/**
+ * Characterization tests for the EntityImp decomposition into EntityImp + UserDialogDelegate.
+ *
+ * <p>Verifies that:
+ * <ul>
+ *   <li>EntityImp's public API surface is preserved (all expected public methods exist)</li>
+ *   <li>UserDialogDelegate is package-private (not public)</li>
+ *   <li>EntityImp is under the target line count</li>
+ *   <li>Dialog delegation methods delegate correctly (signature match)</li>
+ * </ul>
+ */
+public class EntityImpStructureTest {
+
+  private static final Set<String> EXPECTED_PUBLIC_METHODS = Set.of(
+      "getInstance",
+      "getAbstractionFromSgElement",
+      "getName",
+      "setName",
+      "getPropertyForAbstractionGetter",
+      "getAxisAlignedMinimumBoundingBox",
+      "getCollisionHull",
+      "getDynamicAxisAlignedMinimumBoundingBox",
+      "getAbstraction",
+      "getSgComposite",
+      "getSgReferenceFrame",
+      "getActualEntityImplementation",
+      "getVehicle",
+      "setVehicle",
+      "isDescendantOf",
+      "getScene",
+      "getProgram",
+      "getAbsoluteTransformation",
+      "getTransformation",
+      "createStandIn",
+      "createOffsetStandIn",
+      "transformToAwt",
+      "alreadyAdjustedDelay",
+      "delay",
+      "playAudio",
+      "getDoubleFromUser",
+      "getIntegerFromUser",
+      "getBooleanFromUser",
+      "getStringFromUser",
+      "isCollidingWith",
+      "mendSceneGraphIfNecessary",
+      "toString"
+  );
+
+  @Test
+  public void entityImpPublicApiIsPreserved() {
+    Set<String> actual = Arrays.stream(EntityImp.class.getDeclaredMethods())
+        .filter(m -> Modifier.isPublic(m.getModifiers()))
+        .map(Method::getName)
+        .collect(Collectors.toCollection(TreeSet::new));
+
+    for (String expected : EXPECTED_PUBLIC_METHODS) {
+      assertTrue("Missing public method: " + expected, actual.contains(expected));
+    }
+  }
+
+  @Test
+  public void userDialogDelegateIsPackagePrivate() {
+    int modifiers = UserDialogDelegate.class.getModifiers();
+    assertFalse("UserDialogDelegate must not be public", Modifier.isPublic(modifiers));
+    assertFalse("UserDialogDelegate must not be protected", Modifier.isProtected(modifiers));
+    assertFalse("UserDialogDelegate must not be private", Modifier.isPrivate(modifiers));
+  }
+
+  @Test
+  public void userDialogDelegateHasExpectedMethods() {
+    Set<String> dialogMethods = Set.of(
+        "getDoubleFromUser",
+        "getIntegerFromUser",
+        "getBooleanFromUser",
+        "getStringFromUser"
+    );
+
+    Set<String> actual = Arrays.stream(UserDialogDelegate.class.getDeclaredMethods())
+        .filter(m -> Modifier.isPublic(m.getModifiers()))
+        .map(Method::getName)
+        .collect(Collectors.toSet());
+
+    for (String expected : dialogMethods) {
+      assertTrue("UserDialogDelegate missing method: " + expected, actual.contains(expected));
+    }
+  }
+
+  @Test
+  public void dialogMethodSignaturesMatch() throws NoSuchMethodException {
+    // Verify EntityImp delegation methods match delegate signatures
+    Method entityDouble = EntityImp.class.getMethod("getDoubleFromUser", String.class);
+    Method delegateDouble = UserDialogDelegate.class.getMethod("getDoubleFromUser", String.class);
+    assertEquals(entityDouble.getReturnType(), delegateDouble.getReturnType());
+
+    Method entityInt = EntityImp.class.getMethod("getIntegerFromUser", String.class);
+    Method delegateInt = UserDialogDelegate.class.getMethod("getIntegerFromUser", String.class);
+    assertEquals(entityInt.getReturnType(), delegateInt.getReturnType());
+
+    Method entityBool = EntityImp.class.getMethod("getBooleanFromUser", String.class);
+    Method delegateBool = UserDialogDelegate.class.getMethod("getBooleanFromUser", String.class);
+    assertEquals(entityBool.getReturnType(), delegateBool.getReturnType());
+
+    Method entityStr = EntityImp.class.getMethod("getStringFromUser", String.class);
+    Method delegateStr = UserDialogDelegate.class.getMethod("getStringFromUser", String.class);
+    assertEquals(entityStr.getReturnType(), delegateStr.getReturnType());
+  }
+
+  @Test
+  public void entityImpIsUnder500Lines() throws Exception {
+    // Verify via source resource isn't practical, so verify structurally:
+    // EntityImp should have fewer declared methods after extraction
+    long declaredMethodCount = Arrays.stream(EntityImp.class.getDeclaredMethods())
+        .filter(m -> !m.isSynthetic())
+        .count();
+    // After extraction, EntityImp should have significantly fewer methods
+    // than the original 86. The dialog inner classes alone had ~30+ methods.
+    assertTrue("EntityImp should have fewer than 50 declared methods after extraction, got " + declaredMethodCount,
+        declaredMethodCount < 50);
+  }
+}
